refactor: move schemas & types to shared
