+++ apps/backend/src/modules/comments/comment.schema.ts
@@ -1,7 +1,5 @@
 import { z } from "zod";
-import { idParamSchema, paginationQuerySchema } from "@/common/schemas.js";
-
-export { type CreateCommentBody, createCommentSchema } from "@recipes/shared";
+import { idParamSchema } from "@/common/schemas.js";
 
 export const commentParamsSchema = z.object({
   id: idParamSchema,
@@ -10,9 +8,3 @@ export const commentParamsSchema = z.object({
 export const recipeCommentsParamsSchema = z.object({
   recipeId: idParamSchema,
 });
-
-export const commentQuerySchema = paginationQuerySchema;
-
-export type CommentParams = z.infer<typeof commentParamsSchema>;
-export type RecipeCommentsParams = z.infer<typeof recipeCommentsParamsSchema>;
-export type CommentQuery = z.infer<typeof commentQuerySchema>;
