+++ apps/backend/src/modules/favorites/favorite.schema.ts
@@ -1,11 +1,6 @@
 import z from "zod";
-import { idParamSchema, paginationQuerySchema } from "@/common/schemas.js";
+import { idParamSchema } from "@/common/schemas.js";
 
 export const favoriteParamsSchema = z.object({
   recipeId: idParamSchema,
 });
-
-export const favoriteQuerySchema = paginationQuerySchema;
-
-export type FavoriteParams = z.infer<typeof favoriteParamsSchema>;
-export type FavoriteQuery = z.infer<typeof favoriteQuerySchema>;
