+++ apps/backend/src/modules/recipes/recipe.service.ts
@@ -1,4 +1,10 @@
-import type { Paginated, Recipe } from "@recipes/shared";
+import type {
+  CreateRecipeBody,
+  Paginated,
+  Recipe,
+  SearchRecipeQuery,
+  UpdateRecipeBody,
+} from "@recipes/shared";
 import { withPagination } from "@recipes/shared";
 import { isValidObjectId } from "mongoose";
 import type { CacheService } from "@/common/cache/cache.service.js";
@@ -21,12 +27,7 @@ import type {
   CategoryModelType,
 } from "@/modules/categories/index.js";
 import type { FavoriteModelType } from "@/modules/favorites/index.js";
-import type {
-  CreateRecipeBody,
-  RecipeModelType,
-  SearchRecipeQuery,
-  UpdateRecipeBody,
-} from "@/modules/recipes/index.js";
+import type { RecipeModelType } from "@/modules/recipes/index.js";
 import { recipeCache } from "@/modules/recipes/recipe.cache.js";
 import type { UserDocument, UserModelType } from "@/modules/users/index.js";
