+++ packages/shared/src/comments/comment.types.ts
@@ -1,6 +1,7 @@
 import type { z } from "zod";
 import type {
   commentForRecipeSchema,
+  commentQuerySchema,
   commentSchema,
   createCommentSchema,
 } from "./comment.schema.js";
@@ -8,3 +9,5 @@ import type {
 export type CreateCommentBody = z.infer<typeof createCommentSchema>;
 export type Comment = z.infer<typeof commentSchema>;
 export type CommentForRecipe = z.infer<typeof commentForRecipeSchema>;
+
+export type CommentQuery = z.infer<typeof commentQuerySchema>;
