+++ packages/shared/src/favorites/favorite.schema.ts
@@ -1,0 +1,3 @@
+import { paginationQuerySchema } from "../query.js";
+
+export const favoriteQuerySchema = paginationQuerySchema;
