+++ packages/shared/src/favorites/favorite.types.ts
@@ -1,0 +1,4 @@
+import type { z } from "zod";
+import type { favoriteQuerySchema } from "./favorite.schema.js";
+
+export type FavoriteQuery = z.infer<typeof favoriteQuerySchema>;
